+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Changelog
 
+## 1.3.0 (2026-04-30)
+
+- feat: add `inject-async-in-injection-context` rule, and include it in the recommended preset
+
 ## 1.2.0 (2026-04-23)
 
 - feat: support new `@Service()` decorator in Angular 22
 
 ## 1.1.0 (2026-04-16)
 
-- feat: add custom-function-in-injection-context rule
+- feat: add `custom-function-in-injection-context` rule
 
 ## 1.0.2 (2026-04-15)